refactor(stella-cli): split command_deck.rs and agent.rs for the file-size ratchet
Both files exceeded their file-size ratchet allowlist entries after recent
additions, breaking main's CI gate (the `sizes` step in `make gate` runs
before `lint`/`test`, so it masks whatever comes after it).

- command_deck.rs (4710 lines, limit 4693): extract the SKILLS-tab helper
  cluster (deck_slash_commands, skills_snapshot, npx-output parsing,
  install/preview/create-via-LLM, handle_skills_input, and their private
  helpers — ~430 lines) into command_deck/skills.rs. Items the parent
  (command_deck.rs) or the sibling test module (command_deck/tests.rs)
  call are marked pub(super); everything else stays private to the new
  module.
- agent.rs (2235 lines, limit 2211): extract the code-graph build cluster
  (build_code_graph, index_workspace_graph_blocking, GraphSummary,
  format_graph_stats, SessionGraph, spawn_session_graph — ~210 lines) into
  agent/graph.rs, following the same pattern as the existing
  agent/{engine,goal,outcome,prompt,tools}.rs split. build_code_graph and
  spawn_session_graph are re-exported for agent.rs's own use and the
  crate-wide call sites (command_deck.rs, agent/goal.rs); the
  test-only items are imported under #[cfg(test)] to avoid an
  unused-import warning in non-test builds.

Pure mechanical relocation — no behavior change, no witness test needed
(per AGENTS.md). Both ratchet entries are lowered to lock in the gain:
agent.rs 2161 -> 2032, command_deck.rs 4643 -> 4282 (reflecting the final
line counts once this commit and the companion compile-fix commit both
land — see the next commit for why agent.rs's own diff here nets fewer
lines removed than the extraction alone: this commit carries only the
ratchet-relevant hunks, split out of the same file's compile-fix hunk via
targeted `git apply --cached`).

Part of the co-landing ratchet-union fix (see #270).
