+++ scripts/file-size-ratchet.txt
@@ -1,8 +1,8 @@
 # file-size-ratchet allowlist — see scripts/check-file-sizes.sh.
 # One entry per line: <repo-relative path> <line count when recorded>.
 # Entries may only shrink or disappear; never raise a count.
-stella-cli/src/agent.rs 2161
-stella-cli/src/command_deck.rs 4643
+stella-cli/src/agent.rs 2032
+stella-cli/src/command_deck.rs 4282
 stella-context/src/store.rs 1677
 stella-core/src/bus.rs 1671
 stella-core/src/driver.rs 3149
